requirements.txt
# Minimal dependencies for equity momentum rotation backtest
# Python 3.9+ compatible

numpy>=1.20.0,<2.0.0
pandas>=1.3.0,<3.0.0
PyYAML>=5.4.0,<7.0.0
alpaca-py>=0.28.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pytz>=2024.1

# Testing (optional, for development)
pytest>=7.0.0,<9.0.0

# Code quality (optional, for development)
black>=22.0.0,<25.0.0
ruff>=0.1.0,<1.0.0
